created charting reducer
