+++ client/store/spotify.js
@@ -4,8 +4,6 @@ import getRandomSearch, {randomOffset} from './getRandom'
 const GET_ALBUM = 'GET_ALBUM'
 const GET_PLAYLIST = 'GET_PLAYLIST'
 const GET_RANDOM_PLAYLIST = 'GET_RANDOM_PLAYLIST'
-const GET_AUDIO_FEAT = 'GET_AUDIO_FEAT'
-const GET_AUDIO_FEAT_PLAYER = 'GET_AUDIO_FEAT_PLAYER'
 
 const getRandomPlaylist = rPlaylist => ({
   type: GET_RANDOM_PLAYLIST,
@@ -21,56 +19,6 @@ const getUserPlaylist = playlist => ({
   type: GET_PLAYLIST,
   playlist
 })
-
-const getAudioFeat = featArr => ({
-  type: GET_AUDIO_FEAT,
-  featArr
-})
-
-const getAudioFeatPlayer = featArrPlayer => ({
-  type: GET_AUDIO_FEAT_PLAYER,
-  featArrPlayer
-})
-
-export const fetchAudioFeatPlayer = (token, trackId) => {
-  return async dispatch => {
-    try {
-      const {data} = await axios({
-        url: 'https://api.spotify.com/v1/audio-features',
-        method: 'get',
-        headers: {
-          Authorization: 'Bearer ' + token
-        },
-        params: {
-          ids: trackId.toString()
-        }
-      })
-      dispatch(getAudioFeatPlayer(data))
-    } catch (error) {
-      console.error(error)
-    }
-  }
-}
-
-export const fetchAudioFeat = (token, trackId) => {
-  return async dispatch => {
-    try {
-      const {data} = await axios({
-        url: '	https://api.spotify.com/v1/audio-features',
-        method: 'get',
-        headers: {
-          Authorization: 'Bearer ' + token
-        },
-        params: {
-          ids: trackId.toString()
-        }
-      })
-      dispatch(getAudioFeat(data))
-    } catch (error) {
-      console.error(error)
-    }
-  }
-}
 
 export const fetchRPlaylist = token => {
   return async dispatch => {
@@ -146,10 +94,6 @@ export default function(state = initialState, action) {
       return {...state, playlist: action.playlist}
     case GET_RANDOM_PLAYLIST:
       return {...state, rPlaylist: action.rPlaylist}
-    case GET_AUDIO_FEAT:
-      return {...state, featArr: action.featArr}
-    case GET_AUDIO_FEAT_PLAYER:
-      return [...state, action.featArrPlayer]
     default:
       return state
   }
